refactor: backpressure config can build a limiter
