+++ zeebe/broker/src/main/java/io/camunda/zeebe/broker/system/configuration/backpressure/BackpressureCfg.java
@@ -7,7 +7,14 @@
  */
 package io.camunda.zeebe.broker.system.configuration.backpressure;
 
+import com.netflix.concurrency.limits.limit.AbstractLimit;
+import com.netflix.concurrency.limits.limit.FixedLimit;
+import com.netflix.concurrency.limits.limit.Gradient2Limit;
+import com.netflix.concurrency.limits.limit.GradientLimit;
+import com.netflix.concurrency.limits.limit.VegasLimit;
 import io.camunda.zeebe.broker.system.configuration.ConfigurationEntry;
+import io.camunda.zeebe.broker.transport.backpressure.StabilizingAIMDLimit;
+import java.util.concurrent.TimeUnit;
 
 public final class BackpressureCfg implements ConfigurationEntry {
 
@@ -86,6 +93,52 @@ public final class BackpressureCfg implements ConfigurationEntry {
         + ", gradient2="
         + gradient2
         + '}';
+  }
+
+  public AbstractLimit buildLimit() {
+    final LimitAlgorithm algorithm = getAlgorithm();
+    return switch (algorithm) {
+      case AIMD -> getAIMD(getAimd());
+      case FIXED -> FixedLimit.of(getFixed().getLimit());
+      case GRADIENT -> getGradientLimit(getGradient());
+      case GRADIENT2 -> getGradient2Limit(getGradient2());
+      case VEGAS -> getVegasLimit(getVegas());
+    };
+  }
+
+  private static VegasLimit getVegasLimit(final VegasCfg vegasCfg) {
+    return VegasLimit.newBuilder()
+        .alpha(vegasCfg.getAlpha())
+        .beta(vegasCfg.getBeta())
+        .initialLimit(vegasCfg.getInitialLimit())
+        .build();
+  }
+
+  private static Gradient2Limit getGradient2Limit(final Gradient2Cfg gradient2Cfg) {
+    return Gradient2Limit.newBuilder()
+        .rttTolerance(gradient2Cfg.getRttTolerance())
+        .initialLimit(gradient2Cfg.getInitialLimit())
+        .minLimit(gradient2Cfg.getMinLimit())
+        .longWindow(gradient2Cfg.getLongWindow())
+        .build();
+  }
+
+  private static GradientLimit getGradientLimit(final GradientCfg gradientCfg) {
+    return GradientLimit.newBuilder()
+        .minLimit(gradientCfg.getMinLimit())
+        .initialLimit(gradientCfg.getInitialLimit())
+        .rttTolerance(gradientCfg.getRttTolerance())
+        .build();
+  }
+
+  private static StabilizingAIMDLimit getAIMD(final AIMDCfg aimdCfg) {
+    return StabilizingAIMDLimit.newBuilder()
+        .initialLimit(aimdCfg.getInitialLimit())
+        .minLimit(aimdCfg.getMinLimit())
+        .maxLimit(aimdCfg.getMaxLimit())
+        .expectedRTT(aimdCfg.getRequestTimeout().toMillis(), TimeUnit.MILLISECONDS)
+        .backoffRatio(aimdCfg.getBackoffRatio())
+        .build();
   }
 
   public enum LimitAlgorithm {
